data/rag_pipeline/requirements.txt
gitpython
sentence-transformers
faiss-cpu
numpy
